django-backend/requirements.txt
asgiref==3.11.1
click==8.3.2
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.3.0
h11==0.16.0
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
uvicorn==0.44.0
whitenoise==6.12.0
